Cookies refactoring
