--- api/src/main/java/org/asynchttpclient/Cookie.java
@@ -1,401 +1,0 @@
-/*
- * Copyright 2010 Ning, Inc.
- *
- * Ning licenses this file to you under the Apache License, version 2.0
- * (the "License"); you may not use this file except in compliance with the
- * License.  You may obtain a copy of the License at:
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- */
-package org.asynchttpclient;
-
-import java.util.Collections;
-import java.util.Set;
-import java.util.TreeSet;
-
-public class Cookie implements Comparable<Cookie> {
-
-    public static Cookie newValidCookie(String domain, String name, String value, String path, int maxAge, boolean secure) {
-        return newValidCookie(domain, name, value, path, maxAge, secure, 1);
-    }
-
-    public static Cookie newValidCookie(String domain, String name, String value, String path, int maxAge, boolean secure, int version) {
-        return newValidCookie(domain, name, value, value, path, maxAge, secure, version, false, false, null, null, Collections.<Integer> emptySet());
-    }
-
-    public static Cookie newValidCookie(String domain, String name, String value, String rawValue, String path, int maxAge, boolean secure, int version, boolean httpOnly,
-            boolean discard, String comment, String commentUrl, Set<Integer> ports) {
-
-        if (name == null) {
-            throw new NullPointerException("name");
-        }
-        name = name.trim();
-        if (name.isEmpty()) {
-            throw new IllegalArgumentException("empty name");
-        }
-
-        for (int i = 0; i < name.length(); i++) {
-            char c = name.charAt(i);
-            if (c > 127) {
-                throw new IllegalArgumentException("name contains non-ascii character: " + name);
-            }
-
-            // Check prohibited characters.
-            switch (c) {
-            case '\t':
-            case '\n':
-            case 0x0b:
-            case '\f':
-            case '\r':
-            case ' ':
-            case ',':
-            case ';':
-            case '=':
-                throw new IllegalArgumentException("name contains one of the following prohibited characters: " + "=,; \\t\\r\\n\\v\\f: " + name);
-            }
-        }
-
-        if (name.charAt(0) == '$') {
-            throw new IllegalArgumentException("name starting with '$' not allowed: " + name);
-        }
-
-        if (value == null) {
-            throw new NullPointerException("value");
-        }
-
-        domain = validateValue("domain", domain);
-        path = validateValue("path", path);
-
-        if (version > 0) {
-            comment = validateValue("comment", comment);
-        } else {
-            comment = null;
-        }
-        if (version > 1) {
-            commentUrl = validateValue("commentUrl", commentUrl);
-            Set<Integer> newPorts = new TreeSet<Integer>();
-            for (int p : ports) {
-                if (p <= 0 || p > 65535) {
-                    throw new IllegalArgumentException("port out of range: " + p);
-                }
-                newPorts.add(Integer.valueOf(p));
-            }
-            if (newPorts.isEmpty()) {
-                ports = Collections.emptySet();
-            } else {
-                ports = Collections.unmodifiableSet(newPorts);
-            }
-        } else {
-            discard = false;
-            commentUrl = null;
-        }
-
-        return new Cookie(domain, name, value, rawValue, path, maxAge, secure, version, httpOnly, discard, comment, commentUrl, ports);
-    }
-
-    private static String validateValue(String name, String value) {
-        if (value == null) {
-            return null;
-        }
-        value = value.trim();
-        if (value.isEmpty()) {
-            return null;
-        }
-
-        for (int i = 0; i < value.length(); i++) {
-            char c = value.charAt(i);
-            switch (c) {
-            case '\r':
-            case '\n':
-            case '\f':
-            case 0x0b:
-            case ';':
-                throw new IllegalArgumentException(name + " contains one of the following prohibited characters: " + ";\\r\\n\\f\\v (" + value + ')');
-            }
-        }
-        return value;
-    }
-
-    private final String domain;
-    private final String name;
-    private final String value;
-    private final String rawValue;
-    private final String path;
-    private final int maxAge;
-    private final boolean secure;
-    private final int version;
-    private final boolean httpOnly;
-    private final boolean discard;
-    private final String comment;
-    private final String commentUrl;
-    private final Set<Integer> ports;
-
-    public Cookie(String domain, String name, String value, String rawValue, String path, int maxAge, boolean secure, int version, boolean httpOnly, boolean discard,
-            String comment, String commentUrl, Set<Integer> ports) {
-        this.domain = domain;
-        this.name = name;
-        this.value = value;
-        this.rawValue = rawValue;
-        this.path = path;
-        this.maxAge = maxAge;
-        this.secure = secure;
-        this.version = version;
-        this.httpOnly = httpOnly;
-        this.discard = discard;
-        this.comment = comment;
-        this.commentUrl = commentUrl;
-        this.ports = ports;
-    }
-
-    public String getDomain() {
-        return domain;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getValue() {
-        return value;
-    }
-
-    public String getRawValue() {
-        return rawValue;
-    }
-
-    public String getPath() {
-        return path;
-    }
-
-    public int getMaxAge() {
-        return maxAge;
-    }
-
-    public boolean isSecure() {
-        return secure;
-    }
-
-    public int getVersion() {
-        return version;
-    }
-
-    public String getComment() {
-        return this.comment;
-    }
-
-    public String getCommentUrl() {
-        return this.commentUrl;
-    }
-
-    public boolean isHttpOnly() {
-        return httpOnly;
-    }
-
-    public boolean isDiscard() {
-        return discard;
-    }
-
-    public Set<Integer> getPorts() {
-        return ports;
-    }
-
-    @Override
-    public String toString() {
-        StringBuilder buf = new StringBuilder();
-        buf.append(name);
-        buf.append("=");
-        buf.append(rawValue);
-        if (getDomain() != null) {
-            buf.append("; domain=");
-            buf.append(domain);
-        }
-        if (getPath() != null) {
-            buf.append("; path=");
-            buf.append(path);
-        }
-        if (getComment() != null) {
-            buf.append("; comment=");
-            buf.append(comment);
-        }
-        if (getMaxAge() >= 0) {
-            buf.append("; maxAge=");
-            buf.append(maxAge);
-            buf.append("s");
-        }
-        if (secure) {
-            buf.append("; secure");
-        }
-        if (httpOnly) {
-            buf.append("; HTTPOnly");
-        }
-        return buf.toString();
-    }
-
-    public int compareTo(Cookie c) {
-        int v;
-        v = getName().compareToIgnoreCase(c.getName());
-        if (v != 0) {
-            return v;
-        }
-
-        if (getPath() == null) {
-            if (c.getPath() != null) {
-                return -1;
-            }
-        } else if (c.getPath() == null) {
-            return 1;
-        } else {
-            v = getPath().compareTo(c.getPath());
-            if (v != 0) {
-                return v;
-            }
-        }
-
-        if (getDomain() == null) {
-            if (c.getDomain() != null) {
-                return -1;
-            }
-        } else if (c.getDomain() == null) {
-            return 1;
-        } else {
-            v = getDomain().compareToIgnoreCase(c.getDomain());
-            return v;
-        }
-
-        return 0;
-    }
-
-    public static class CookieBuilder {
-
-        private final String name;
-        private final String value;
-        private final String rawValue;
-        private String domain;
-        private String path;
-        private int maxAge = -1;
-        private boolean secure;
-        private int version;
-        private boolean httpOnly;
-        private boolean discard;
-        private String comment;
-        private String commentUrl;
-        private Set<Integer> ports;
-        private boolean domainNotSet = true;
-        private boolean pathNotSet = true;
-        private boolean maxAgeNotSet = true;
-        private boolean secureNotSet = true;
-        private boolean versionNotSet = true;
-        private boolean httpOnlyNotSet = true;
-        private boolean discardNotSet = true;
-        private boolean commentNotSet = true;
-        private boolean commentUrlNotSet = true;
-        private boolean portsNotSet = true;
-
-        public CookieBuilder(String name, String value, String rawValue) {
-            this.name = name;
-            this.value = value;
-            this.rawValue = rawValue;
-        }
-
-        public Cookie build() {
-            return new Cookie(domain, name, value, rawValue, path, maxAge, secure, version, httpOnly, discard, comment, commentUrl, ports);
-        }
-
-        public void setDomain(String domain) {
-            this.domain = domain;
-            domainNotSet = false;
-        }
-
-        public void setPath(String path) {
-            this.path = path;
-            pathNotSet = false;
-        }
-
-        public void setMaxAge(int maxAge) {
-            this.maxAge = maxAge;
-            maxAgeNotSet = false;
-        }
-
-        public void setSecure(boolean secure) {
-            this.secure = secure;
-            secureNotSet = false;
-        }
-
-        public void setVersion(int version) {
-            this.version = version;
-            versionNotSet = false;
-        }
-
-        public void setHttpOnly(boolean httpOnly) {
-            this.httpOnly = httpOnly;
-            httpOnlyNotSet = false;
-        }
-
-        public void setDiscard(boolean discard) {
-            this.discard = discard;
-            discardNotSet = false;
-        }
-
-        public void setComment(String comment) {
-            this.comment = comment;
-            commentNotSet = false;
-        }
-
-        public void setCommentUrl(String commentUrl) {
-            this.commentUrl = commentUrl;
-            commentNotSet = false;
-        }
-
-        public void setPorts(Set<Integer> ports) {
-            this.ports = ports;
-            portsNotSet = false;
-        }
-
-        public boolean isDomainNotSet() {
-            return domainNotSet;
-        }
-
-        public boolean isPathNotSet() {
-            return pathNotSet;
-        }
-
-        public boolean isMaxAgeNotSet() {
-            return maxAgeNotSet;
-        }
-
-        public boolean isSecureNotSet() {
-            return secureNotSet;
-        }
-
-        public boolean isVersionNotSet() {
-            return versionNotSet;
-        }
-
-        public boolean isHttpOnlyNotSet() {
-            return httpOnlyNotSet;
-        }
-
-        public boolean isDiscardNotSet() {
-            return discardNotSet;
-        }
-
-        public boolean isCommentNotSet() {
-            return commentNotSet;
-        }
-
-        public boolean isCommentUrlNotSet() {
-            return commentUrlNotSet;
-        }
-
-        public boolean isPortsNotSet() {
-            return portsNotSet;
-        }
-    }
-}
