+++ api/src/main/java/org/asynchttpclient/cookie/CookieEncoder.java
@@ -1,0 +1,47 @@
+/*
+ * Copyright (c) 2014 AsyncHttpClient Project. All rights reserved.
+ *
+ * This program is licensed to you under the Apache License Version 2.0,
+ * and you may not use this file except in compliance with the Apache License Version 2.0.
+ * You may obtain a copy of the Apache License Version 2.0 at http://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the Apache License Version 2.0 is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the Apache License Version 2.0 for the specific language governing permissions and limitations there under.
+ */
+package org.asynchttpclient.cookie;
+
+import java.util.Collection;
+
+public final class CookieEncoder {
+
+    private CookieEncoder() {
+    }
+
+    public static String encodeClientSide(Collection<Cookie> cookies) {
+        StringBuilder sb = new StringBuilder();
+
+        for (Cookie cookie : cookies) {
+            add(sb, cookie.getName(), cookie.getRawValue());
+        }
+
+        if (sb.length() > 0) {
+            sb.setLength(sb.length() - 2);
+        }
+        return sb.toString();
+    }
+
+    private static void add(StringBuilder sb, String name, String val) {
+
+        if (val == null) {
+            val = "";
+        }
+
+        sb.append(name);
+        sb.append('=');
+        sb.append(val);
+        sb.append(';');
+        sb.append(' ');
+    }
+}
