--- api/src/main/java/org/asynchttpclient/org/jboss/netty/util/internal/StringUtil.java
@@ -1,72 +1,0 @@
-/*
- * Copyright 2012 The Netty Project
- *
- * The Netty Project licenses this file to you under the Apache License,
- * version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at:
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-package org.asynchttpclient.org.jboss.netty.util.internal;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * String utility class.
- */
-public final class StringUtil {
-
-    private StringUtil() {
-        // Unused.
-    }
-
-    private static final String EMPTY_STRING = "";
-
-    /**
-     * Splits the specified {@link String} with the specified delimiter.  This operation is a simplified and optimized
-     * version of {@link String#split(String)}.
-     */
-    public static String[] split(String value, char delim) {
-        final int end = value.length();
-        final List<String> res = new ArrayList<String>();
-
-        int start = 0;
-        for (int i = 0; i < end; i ++) {
-            if (value.charAt(i) == delim) {
-                if (start == i) {
-                    res.add(EMPTY_STRING);
-                } else {
-                    res.add(value.substring(start, i));
-                }
-                start = i + 1;
-            }
-        }
-
-        if (start == 0) { // If no delimiter was found in the value
-            res.add(value);
-        } else {
-            if (start != end) {
-                // Add the last element if it's not empty.
-                res.add(value.substring(start, end));
-            } else {
-                // Truncate trailing empty elements.
-                for (int i = res.size() - 1; i >= 0; i --) {
-                    if (res.get(i).length() == 0) {
-                        res.remove(i);
-                    } else {
-                        break;
-                    }
-                }
-            }
-        }
-
-        return res.toArray(new String[res.size()]);
-    }
-}
