+++ api/src/main/java/org/asynchttpclient/providers/ResponseBase.java
@@ -9,12 +9,12 @@ import java.nio.ByteBuffer;
 import java.util.Collections;
 import java.util.List;
 
-import org.asynchttpclient.Cookie;
 import org.asynchttpclient.FluentCaseInsensitiveStringsMap;
 import org.asynchttpclient.HttpResponseBodyPart;
 import org.asynchttpclient.HttpResponseHeaders;
 import org.asynchttpclient.HttpResponseStatus;
 import org.asynchttpclient.Response;
+import org.asynchttpclient.cookie.Cookie;
 import org.asynchttpclient.util.AsyncHttpProviderUtils;
 
 public abstract class ResponseBase implements Response {
