+++ api/src/test/java/org/asynchttpclient/async/AsyncProvidersBasicTest.java
@@ -46,13 +46,13 @@ import org.asynchttpclient.AsyncHttpClientConfig;
 import org.asynchttpclient.AsyncHttpClientConfig.Builder;
 import org.asynchttpclient.AsyncHttpClientConfigBean;
 import org.asynchttpclient.AsyncHttpProviderConfig;
-import org.asynchttpclient.Cookie;
 import org.asynchttpclient.FluentCaseInsensitiveStringsMap;
 import org.asynchttpclient.MaxRedirectException;
 import org.asynchttpclient.ProxyServer;
 import org.asynchttpclient.Request;
 import org.asynchttpclient.RequestBuilder;
 import org.asynchttpclient.Response;
+import org.asynchttpclient.cookie.Cookie;
 import org.asynchttpclient.multipart.Part;
 import org.asynchttpclient.multipart.StringPart;
 import org.asynchttpclient.util.StandardCharsets;
