+++ providers/grizzly/src/main/java/org/asynchttpclient/providers/grizzly/filters/AsyncHttpClientFilter.java
@@ -16,7 +16,6 @@ package org.asynchttpclient.providers.grizzly.filters;
 import org.asynchttpclient.AsyncCompletionHandler;
 import org.asynchttpclient.AsyncHandler;
 import org.asynchttpclient.AsyncHttpClientConfig;
-import org.asynchttpclient.Cookie;
 import org.asynchttpclient.FluentCaseInsensitiveStringsMap;
 import org.asynchttpclient.FluentStringsMap;
 import org.asynchttpclient.ProxyServer;
@@ -24,6 +23,7 @@ import org.asynchttpclient.Request;
 import org.asynchttpclient.RequestBuilder;
 import org.asynchttpclient.Response;
 import org.asynchttpclient.UpgradeHandler;
+import org.asynchttpclient.cookie.Cookie;
 import org.asynchttpclient.listener.TransferCompletionHandler;
 import org.asynchttpclient.providers.grizzly.GrizzlyAsyncHttpProvider;
 import org.asynchttpclient.providers.grizzly.GrizzlyResponseFuture;
@@ -470,7 +470,6 @@ public final class AsyncHttpClientFilter extends BaseFilter {
                 final org.glassfish.grizzly.http.Cookie gCookie = new org.glassfish.grizzly.http.Cookie(cookie.getName(), cookie.getValue());
                 gCookie.setDomain(cookie.getDomain());
                 gCookie.setPath(cookie.getPath());
-                gCookie.setVersion(cookie.getVersion());
                 gCookie.setMaxAge(cookie.getMaxAge());
                 gCookie.setSecure(cookie.isSecure());
                 gCookies[idx] = gCookie;
